Option to export OTLP metrics during benchmarks (#1610)
Adds a flag to export OTLP metrics during benchmarks

With the change, we can test the impact of metrics on Mountpoint
throughput. This change also includes some refactoring that was missed
while removing benchmark config parsing logic in #1609.

### Does this change impact existing behavior?

No, benchmarks only

### Does this change need a changelog entry? Does it require a version
change?

No, benchmarks only

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>
Signed-off-by: Tadiwa Magwenzi <tadiwaom@amazon.com>
Co-authored-by: Tadiwa Magwenzi <tadiwaom@amazon.com>