Add checking step to PUBLISHING_CRATES.md (#1607)
We currently don't have a step for checking the newly published
versions. This change adds links to `crates.io` to facilitate this.

### Does this change impact existing behavior?

No, only change to documentation.

### Does this change need a changelog entry? Does it require a version
change?

No, only change to documentation.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>