Switch to real-time recording for OTLP metrics.  (#1615)
Switch OTLP metrics from batch recording to real-time recording directly
to OTel SDK instruments. This eliminates the need for batch aggregation
of histogram samples. This change also adds support for histogram export
via OTLP. Log-based metrics remain unchanged.

With this change, we are also removing the use of counter.absolute()
method as OTel SDK doesn't provide an equivalent method to set absolute
values for counters. We should use gauge or histogram to record absolute
values.

### Does this change impact existing behavior?

No, the changes are under a feature flag.

### Does this change need a changelog entry? Does it require a version
change?

No, the changes are under a feature flag. 

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>