Use exponential histogram view for histogram metrics.
With explicit hitograms, OTel uses fixed bucket bounds, which don't work
well for all histogram metrics. So we need to manually configure the
bucket bounds for different metrics with different boundaries.

This is not a problem for exponential histograms as the buckets are
automatically adjusted and provide more accurate metrics. However, this
takes a dependency on OTel SDK's spec_unstable_metrics_views feature,
which is not fully stable

Signed-off-by: Sahitya Damera <sahityad@amazon.com>