Implement integrity checks for the metadata store (#1563)
Introduce checksum verification of:

1. entries provided in the input CSV manifest (and for the whole file)
2. entries stored in the metadata store

Slightly refactored `ManifestEntry` in order to avoid huge `match`
blocks. This PR also adds an example of the manifest creation.

### Does this change impact existing behavior?

No, only `mount_from_config` example.

### Does this change need a changelog entry? Does it require a version
change?

Yes, an entry for `mountpoint-s3-fs` crate, and a minor version bump of
this crate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>