Make `mnt::Mount` public (#1098)
fuser v0.15.0 added support for creating sessions from FUSE fd. I'm
working on a PR to add this support to Mountpoint. We still need to open
the FUSE device and call `mount` syscall in order to test this new
behavior, and all this logic is already exists with `mnt::Mount`.

We could just copy the logic from `mnt::Mount` - but since we already
have this fork, making this change seemed fine to me. Maybe we can also
consider upstreaming it.

No

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Burak Varli <burakvar@amazon.co.uk>