Skip fuser tests on macOS in CI (#1636)
Avoid running unit tests in CI for the `mountpoint-s3-fuser` crate on
macOS. Since `fuser` also removed their macOS CI, we don't want to
trigger spurious failures on a platform we do not support.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>