Merge changes from the fuser/fork branch (#1633)
Pull in changes from the `fuser/fork` branch, updating the
`mountpoint-s3-fuser` crate to `v0.1.1`.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>