Update S3 request metrics for OTLP export (#1630)
This PR defines metric configuration for OTLP export and adds support
for passing units and filtering stable metrics for otlp export.

This change also updates S3 request metrics to use the new metric names
and attributes in both logs and OTLP export.

### Does this change impact existing behavior?

Changes log format for S3 request metrics in logs. The rest of the
changes are under the otlp_integration feature flag.

### Does this change need a changelog entry? Does it require a version
change?

Requires a changelog entry as it updates the format of logs.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>