Fix typing in benchmark script (#1646)
`any` isn't a type; also no need for the shortcut, instead specify the
proper return type for the resource monitoring functions.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Andrew Peace <adpeace@amazon.com>