Unstable flag to open a new FD on each fuse thread. (#1642)
This mimics the libfuse behaviour. This is currently behind a flag
because the performance and compatibility concerns are still being
considered.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Andrew Peace <adpeace@amazon.com>