Update S3 log analyzer to parse memory usage correctly (#1652)
We recently updated the log format to include an extra space, which
broke parsing of memory usage metric from the logs. With this change,
log analyzer should be able to parse units or additional spacing before
the metric.

### Does this change impact existing behavior?

No, only updates log-analyzer

### Does this change need a changelog entry? Does it require a version
change?

No, only updates log-analyzer

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Sahitya Damera <sahityad@amazon.com>