Add Docker build (no publish) in CI, fix image sources, minor style changes (#1665)
Building the container images had warnings due to style inconsistencies.
Additionally, the base image did not use the ECR images in all cases. On
top of addressing these two issues, this PR adds a job to CI to verify
that the container images are buildable.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>