Update on_telemetry to use operation_name rather than request_type for metrics (#1669)
Previously, Mountpoint metrics reported quite a lot of S3 operations as
`"Default"` operations. This was due to leaking of the underlying meta
request abstraction.

This change replaces that by using the `operation_name` provided by the
CRT request metrics struct. This should cover all S3 operations we care
about.

### Does this change impact existing behavior?

This change fixes metric attributes. It is not considered a breaking
change, as the metric log format is considered "unstable".

### Does this change need a changelog entry? Does it require a version
change?

No, minor fix to metric attributes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>