Keep a constant memory reservation for backwards seek for each fh (#1631)
Currently we reserve memory for backwards seek only when an actual seek
occurs. The memory is used even if there is no such seek. Also we
reserve too few memory, up to `1MiB`, while the whole extra buffer of
size `part_size` may be kept in RAM.

With this change MP makes a memory reservation upon the creation of
`PrefetchGetObject` and releases memory once it is dropped. This is done
in addition to the existing mechanism which reserves memory in
`PartQueue::push_front`.

### Does this change impact existing behavior?

In a memory constrained environment, this may result in smaller read
window sizes and less memory consumption.

### Does this change need a changelog entry? Does it require a version
change?

Patch version change and a change log to `mountpoint-s3-fs`, will add
later.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Signed-off-by: Vlad Volodkin <vlaad@amazon.co.uk>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.co.uk>