Add units for metrics in logs (#1677)
This change adds units canonical labels to metrics in logs. Currently,
it only includes units for metrics eligible for OTLP export.

### Does this change impact existing behavior?

Yes, it updates the log format

### Does this change need a changelog entry? Does it require a version
change?

Yes, it updates the log format

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Sahitya Damera <sahityad@amazon.com>