Add otlp metrics user-agent tag (#1686)
This adds a user-agent tag when otlp endpoint is configured to get
insights into usage of otlp metrics.

### Does this change impact existing behavior?

No, under a feature flag

### Does this change need a changelog entry? Does it require a version
change?

No. 

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Sahitya Damera <sahityad@amazon.com>