Update RPM workflow to handle configurable release (#1690)
Update the RPM workflow to use a "version tag" (`<VERSION>-<RELEASE>`),
which will allow to handle a configurable release in a future change.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>
Co-authored-by: Tadiwa Magwenzi <tadiwaom@amazon.com>