Use explicit result type in readdir reply callback (#1700)
The readdir reply callback in the `metablock` module was misleading and
its documentation was wrong: `TryAddDirEntry` would return `false` on
success (contrary to the rustdoc and common sense). Note though that all
existing invocations were using it correctly.

This change introduces an explicit return type to avoid any confusion,
at the cost of some increase in verbosity.

### Does this change impact existing behavior?

No, code tidy only.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>