Add 'workflow complete' GHA targets to codify CI requirements (#1702)
Mountpoint's GitHub repository today blocks PR approvals based on a list
of jobs in the branch protection rules. This is not visible to
contributors and also not possible to update unless you have additional
privileges on the repository. When new workflows are added, it is
frequently missed to update the branch protection rules.

This change instead defines targets within the workflow YAML files, such
that it is clear what the requirements are. Changes in requirements can
go through version control and code review.

I've only covered the tests and integration test workflows, as most
other workflows have only one job. These workflows are also where most
changes occur.

Following this commit, the GitHub branch protection rules should be
updated to require only these new targets for these two workflows.

### Does this change impact existing behavior?

It changes CI only. There will be no actual change in behavior, just a
change in how we maintain the rules.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>