Add AL2023 to installation documentation (#1705)
Updating Installation documentation to account for AL2023.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No. It is a documentation update

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Tadiwa Magwenzi <tadiwaom@amazon.com>