Tidy up propagation of remote status from the superblock (#1709)
Internal change to correctly propagate remote status information from
the internal inode representation in the superblock to the consuming
file system. This fixes a gap introduced with the `Metablock`
abstraction (#1500), where the remote status was checked separately from
the inode's `stat`.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>