Fix client-bp benchmarks (#1711)
Fixed a problem that occurred in client backpressure benchmarks.

When the benchmark was of client backpressure type, the code was
checking the `client` config for the `read_window_size` metric and not
the `client_bp` config. Thus it was erroring out as client does not have
a read window size.


### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No, Bug fix.
---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Tadiwa Magwenzi <tadiwaom@amazon.com>