Improve CONTRIBUTING.md (#1676)
In order to improve contribution experience, this PR:
- Creates the `make pre-pr-check` command
- Adjusts the documentation link in PR template to the full PR section
instead of just a subsection
- Improves `CONTRIBUTING.md` with small adjustments

### Does this change impact existing behavior?

No, this is basically only changes to documentation.

### Does this change need a changelog entry? Does it require a version
change?

No, this is basically only changes to documentation.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Renan Magagnin <renanmag@amazon.co.uk>
Signed-off-by: Renan Magagnin <renanmagagnin@gmail.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>