Fix workflow permissions to publish benchmark results (#1722)
The changes in #1695 resulted in benchmark actions on push to main to
fail when trying to publish results. This change will allow benchmark
workflows to write to github pages by fixing their permissions.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>