Fix internal race condition in the incremental upload queue (#1728)
This change addresses an issue that caused the
`upload::incremental::tests::test_append_failure_on_object_replaced` to
fail
[occasionally](https://github.com/awslabs/mountpoint-s3/actions/runs/20229391488/job/58068496627#step:8:576).
The root cause was a race condition in `AppendUploadQueue` when
attempting to write more data after a failure to retrieve the checksum
algorithm, which would occasionally result in a panic instead of
returning an `UploadAlreadyTerminated` error.

The issue could be fixed with a simple change (see first commit), but I
opted to refactor the queue to use a single channel to return both the
checksum algorithm and the PutObject responses, making error handling
more uniform.

**Note that there is no user impact**, since further writes after an
error are not allowed by the "file system" layer.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>