Release crates, including mountpoint-s3-fs 0.8.4 (#1734)
Bump versions / update changelogs of the following crates prior to the
release:

- `mountpoint-s3-crt`
- `mountpoint-s3-client`
- `mountpoint-s3-fs`

Diff with the previous release:
https://github.com/awslabs/mountpoint-s3/compare/mountpoint-s3-fs-0.8.3..main

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.co.uk>
Co-authored-by: Vlad Volodkin <vlaad@amazon.co.uk>