Fix internal failure on atomic upload (#1733)
Improve handling of errors on `CreateMultiPartUpload` in the atomic
upload code path. Similarly to the change in #1728, the issue only
manifests when attempting to further write or complete an upload after
an error and it does not affect Mountpoint file system users, since
that's already prevented at that level.

### Does this change impact existing behavior?

No, user-visible behavior not impacted.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>